upgrade pyo3

+++ pyanndata/src/data/slice.rs
@@ -18,7 +18,7 @@ pub fn to_select_info(ob: &Bound<'_, PyAny>, shape: &Shape) -> PyResult<SelectIn
 
 pub fn to_select_elem(ob: &Bound<'_, PyAny>, length: usize) -> PyResult<SelectInfoElem> {
     let select = if let Ok(slice) = ob.downcast::<pyo3::types::PySlice>() {
-        let s = slice.indices(length as i64)?;
+        let s = slice.indices(length as isize)?;
         ndarray::Slice { 
             start: s.start,
             end: Some(s.stop),
